Adding progress in case optimal actor seelection
The added code allows analysis of trace files to generate a
deterministic actor selection schedule for exactly the mapping and input
used to generate the trace. This schedule is intended to represent the
baseline of any "smart" scheduler, since we compile the precomputed
decisions into the code and the only schedule overhead is due to cache
misses.

Compilation using this schedule is currently only implemented in the correspondingly
modified exelixi CPP backend.

The schedule generation currently uses a run length encoding
+ modified Ktails algorithm to generate the state machine,
which runs in n^2 over the length of the input trace.
For nonpreemptive actor selection this runs in acceptable time on the
normal test traces.

Multiple partitions are *somewhat* supported, but currently use hot looping to
verify the scheduled actor fired (i.e. we wait for data to arrive if the
input data is not there yet), and this test can fail if part of the data
arrives with a delay when compared to the originally used trace.